# openapi/src/main/resources/openapi.yml
swagger: '2.0'
info:
  title: Portfolio API
  version: '1.0.0'
  contact:
    name: Alexander Fisher
    email: afisher@alexander-fisher.com
paths:
...
  /projects/{projectId}/gallery:
    get:
      operationId: getProjectGallery
      description: Returns project gallery data for a project
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/ProjectId'
      responses:
        '200':
          description: Project gallery data for a project
          schema:
            $ref: '#/definitions/ProjectGallery'
parameters:
...
  ProjectId:
    name: projectId
    in: path
    type: integer
    required: true
definitions:
...
  Gallery:
    type: object
    properties:
      id:
        type: integer
      coverId:
        type: integer
      title:
        type: string
      slug:
        type: string
  Image:
    type: object
    properties:
      id:
        type: integer
      galleryId:
        type: integer
      filename:
        type: string
      alt:
        type: string
...
  ProjectGallery:
    type: object
    properties:
      gallery:
        $ref: '#/definitions/Gallery'
      images:
        type: array
        items:
          $ref: '#/definitions/Image'